refactor(storage)!: rename storage_providers back to storage
Use the singular 'storage' (mass noun) rather than 'storage_providers'.
Aligns with Django/LangChain/web3.py conventions and matches how 'storage'
reads naturally in English. wallets/ stays plural (count noun) and utils/
stays plural (Python idiom) — each follows its natural English form, and
mixing is standard practice in the Python ecosystem.

Pure rename: 15 files updated, no behaviour change. 343/343 tests pass.
