+++ tests/README.md
@@ -35,6 +35,6 @@ pytest -v
 | `test_paymaster.py` | `Paymaster` — RPC helpers, sponsorability checks |
 | `test_local_storage.py` | `LocalStorageProvider` — file I/O, permissions, path traversal |
 | `test_ipfs_storage.py` | `IPFSStorageProvider` — Pinata upload, CID validation, gateway |
-| `test_storage_providers_from_env.py` | `LocalStorageProvider.from_env` / `IPFSStorageProvider.from_env` — env config per provider |
+| `test_storage_from_env.py` | `LocalStorageProvider.from_env` / `IPFSStorageProvider.from_env` — env config per provider |
 | `test_module_system.py` | `BNBAgentModule` / `ModuleRegistry` — module composition |
 | `test_multicall.py` | `multicall_read` — Multicall3 batch reads |
