+++ tests/test_apex_job_ops.py
@@ -173,7 +173,7 @@ class TestVerifyJob:
 class TestSubmitResult:
     @pytest.mark.asyncio
     async def test_submit_uploads_and_returns_deliverable(self, tmp_path):
-        from bnbagent.storage_providers.local_provider import LocalStorageProvider
+        from bnbagent.storage.local_provider import LocalStorageProvider
 
         storage = LocalStorageProvider(str(tmp_path))
         ops = _make_ops(storage=storage, agent_url="http://agent.example/apex")
@@ -225,7 +225,7 @@ class TestSubmitResult:
 
     @pytest.mark.asyncio
     async def test_within_caps_proceeds(self, tmp_path):
-        from bnbagent.storage_providers.local_provider import LocalStorageProvider
+        from bnbagent.storage.local_provider import LocalStorageProvider
 
         storage = LocalStorageProvider(str(tmp_path))
         ops = _make_ops(storage=storage, agent_url="http://agent.example/apex")
@@ -242,7 +242,7 @@ class TestSubmitResult:
 
     @pytest.mark.asyncio
     async def test_file_url_rewritten_to_agent_endpoint(self, tmp_path):
-        from bnbagent.storage_providers.local_provider import LocalStorageProvider
+        from bnbagent.storage.local_provider import LocalStorageProvider
 
         storage = LocalStorageProvider(str(tmp_path))
         ops = _make_ops(storage=storage, agent_url="http://myagent.example/apex")
@@ -285,7 +285,7 @@ class TestSubmitResult:
 
     @pytest.mark.asyncio
     async def test_file_url_without_agent_url_raises(self, tmp_path):
-        from bnbagent.storage_providers.local_provider import LocalStorageProvider
+        from bnbagent.storage.local_provider import LocalStorageProvider
 
         storage = LocalStorageProvider(str(tmp_path))
         ops = _make_ops(storage=storage, agent_url=None)
